pyproject.toml
[project]
name = "py-load-uniprot"
version = "0.1.0"
description = "A high-performance Python package for ETL processing of UniProtKB data into relational databases."
authors = [
    {name = "Gowtham Rao", email = "rao@ohdsi.org"},
]
dependencies = ["requests>=2.32.5", "pydantic>=2.11.7", "typer[all]>=0.17.4", "rich>=14.1.0", "pydantic-settings>=2.10.1", "lxml>=5.2.2", "psycopg2-binary>=2.9.9"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
pytest = "pytest"

[project.scripts]
py-load-uniprot = "py_load_uniprot.cli:main"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-mock>=3.15.0",
]
